+++ linux/flutter/generated_plugins.cmake
@@ -5,6 +5,7 @@
 list(APPEND FLUTTER_PLUGIN_LIST
   flutter_secure_storage_linux
   gtk
+  objectbox_flutter_libs
   url_launcher_linux
 )