+++ windows/flutter/generated_plugins.cmake
@@ -8,7 +8,9 @@ list(APPEND FLUTTER_PLUGIN_LIST
   firebase_core
   flutter_secure_storage_windows
   geolocator_windows
+  objectbox_flutter_libs
   permission_handler_windows
+  share_plus
   url_launcher_windows
 )